requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.8
Pillow>=9.0
python-barcode>=0.15.1
qrcode>=7.4.2
pypdfium2>=4.0
bleak>=0.22
crc8>=0.2.0
pyserial>=3.5
python-lzo>=1.15
pyobjc-framework-IOBluetooth>=10.0; sys_platform == "darwin"
winsdk>=1.0.0b10; sys_platform == "win32"
